Fixes to the LitSolid shader to ensure that lights that don't have "cast shadows" enabled don't contribute their lighting value to the lighting calculation for lightmapped objects, and made sure that the environment map is shaded by the lightmap properly, if both are used.
Also fixed bugs in the DiffLightMap technique. Like DiffLightMapEnvCube, it needs to subtract, not add the lighting, and define the LIGHTMAP shader define for the VS and PS in order for the lighting to be calculated correctly, or again, you have lights contributing light info to lightmapped objects in incorrect ways.