Make string to floating-point conversion independent from locale.
Instead of using strtod() for converting strings to doubles, use strtod_l()
with an explicit "C" locale. The problem of strtod() is that the decimal
separator changes with the locale. So an XML written on an English system
cannot be read on a German system, for example.

Using strtod_l() + "C" locale uses a dot as separator all the time (locale-independent).