Include <xlocale.h> before <stdlib.h>
I think that on macOS <xlocale.h> defines a macro, which is used by <stdlib.h> to
make the extended strtoX_l() functions available. This means that the inclusion
order is important. (+7 squashed commits)
Squashed commits:
[54ddbe6] Include stdlib.h on macOS too.
[cd1845f] Fix the macOS build.

On macOS, one should include <xlocale.h> in addition to <locale.h>.
[a7d757f] Overlooked one strtod() call.
[62a8f53] Include <locale.h> to fix the Windows build.
[f621cda] Fix Windows build.

Windows has a _create_locale() function instead of newlocale().
[65c4164] Fix Windows build.

Windows has _strtod_l(). Note the leading underscore.
[97e54e4] Make string to floating-point conversion independent from locale.

Instead of using strtod() for converting strings to doubles, use strtod_l()
with an explicit "C" locale. The problem of strtod() is that the decimal
separator changes with the locale. So an XML written on an English system
cannot be read on a German system, for example.

Using strtod_l() + "C" locale uses a dot as separator all the time (locale-independent).