Floating point values should be written with a dot as separator.
Floating point values should always use a '.' as decimal separator and not the separator
determined by the user's locale. Otherwise, XML files written on one system with an
English locale cannot be parsed back correctly on a system with e.g. a German locale.

Replacements for strtod() and sprintf() have been collected in Core/Io.{h,cpp}. Since
sprintf() has a variadic parameter list, the replacement Atomic_sprintf() is implemented
as a macro for simplicity.