Keep the locale when initializing the CEF.
On Linux, the Chromium Embedded Framework chooses the locale based on the environment
variables of the system. Calling CefInitialize() causes the locale to change, potentially
changing the decimal separator of floating point numbers.

The solution is to cache the locale and restore it once CEF has been initialized.

I also reverted the workarounds in StringUtils.cpp, which are not needed any longer.